# src/docs/ApiDocsHopeRaiser.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "HopeRaiser",
	 "description": "CRUD API Documentation using Postman & Swagger",
    "contact": {}
  },
  "host": "http://localhost:3001",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users/update/{id}": {
      "put": {
        "summary": "Change Password",
        "tags": [
          "Users"
        ],
        "operationId": "ChangePassword",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "access_token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ChangePasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully Change Password",
            "headers": {}
          },
          "400": {
            "description": "Failed Change Password",
            "headers": {}
          }
        }
      }
    },
    "/users/delete/{id}": {
      "delete": {
        "summary": "Delete User",
        "tags": [
          "Users"
        ],
        "operationId": "DeleteUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "headers": {}
          },
          "400": {
            "description": "Failed Delete User",
            "headers": {}
          },
          "404": {
            "description": "Error Not Found",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ChangePasswordRequest": {
      "title": "ChangePasswordRequest",
      "example": {
        "password": "admin"
      },
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        }
      },
      "required": [
        "password"
      ]
    },
...
  },
...
}